# doc/api.yaml
openapi: 3.0.3
info:
  title: Simplified login API specification
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "1"
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  responses:
    InvalidRequestError:
      description: bad request
      content:
        application/json:
          schema:
            description: '400 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Bad request, retry!" 
    ForbiddenError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            description: '403 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 50
                example: "Access token is missing or invalid!" 
    InternalServerError:
      description: server error
      content:
        application/json:
          schema:
            description: '500 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Server error, try again later!"
...
  schemas:
    GeneralId:
      type: object
      description: id of a user, a conversation, a group, a message or a comment
      properties:
        value:
          type: integer
          description: id value
          minimum: 0
          maximum: 10000
    Username:
      type: object
      description: username
      properties:
        name:
          type: string
          description: | 
            text that represents the username 
          pattern: '^[a-z0-9]{1,15}$'
          minLength: 1
          maxLength: 15
      required: [name]
    Message:
      type: object
      description: message of a conversation
      properties:
        id:
          $ref: '#/components/schemas/GeneralId'
        date_time:
          type: string
          description: date and time of the message
          format: date-time
          pattern: '^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])T([01]\\d|2[0-3]):[0-5]\\d$'
          minLength: 16
          maxLength: 16
        text:
          description: text of the message
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 9999
        status:
          type: boolean
          description: status of the message
        conv_id:
          $ref: '#/components/schemas/GeneralId'
        sender_id:
          $ref: '#/components/schemas/GeneralId'
        photo:
          type: string
          description: photo of the message
          format: byte
          pattern: '^.*?$'
          minLength: 1
          maxLength: 7000000
        gif:
          type: string
          description: gif of the message
          format: byte
          pattern: '^.*?$'
          minLength: 1
          maxLength: 7000000
        forward:
          type: boolean
          description: define if is a forwarded message
        reply_id:
          $ref: '#/components/schemas/GeneralId'
...
    Group:
      type: object
      description: group with other users
      properties:
        id:
          $ref: '#/components/schemas/GeneralId'
        group_name:
          type: string
          description: text that represents the group name
          pattern: '^.*?$'
          minLength: 1
          maxLength: 20
        photo:
          type: string
          description: group photo
          pattern: '^.*?$'
          format: byte
          minLength: 1
          maxLength: 7000000
      required: [id, group_name]
    User:
      type: object
      description: general user
      properties:
        id:
          $ref: '#/components/schemas/GeneralId'
        user_name:
          $ref: '#/components/schemas/Username'
        photo:
          type: string
          description: group photo
          pattern: '^.*?$'
          format: byte
          minLength: 1
          maxLength: 7000000
      required: [id, user_name]
...
paths:
...
  /users/{user}/conversations/{conv_id}/messages/{mess_id}:
    description: message area, where user can forward or delete it
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: conv_id
        description: receiver id (group or user)
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: mess_id
        description: message id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    post:
      security:
        - bearerAuth: []
      tags: ["forwarding of a message"]
      operationId: forwardMessage
      summary: forwards the message
      description: User forwards a message in a chat, represented by his id
      requestBody:
        description: User gives the chat id in input
        content:
          application/json:
            schema:
              type: object
              description: object that contains array of conversations data
              properties:
                destination:
                  description: conversations array in input
                  type: array
                  items:
                    type: object
                    description: object with all data of a conversation
                    properties: 
                      user_id:
                        $ref: '#/components/schemas/GeneralId'
                      group_id:
                        $ref: '#/components/schemas/GeneralId'
                  minItems: 0
                  maxItems: 10000
      responses:
        '201':
          description: message forwarding action successful
          content:
            application/json:
              schema:
                description: '201 response message'
                type: object
                properties:
                  messages:
                    description: '201 response message'
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                    minItems: 0
                    maxItems: 10000
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      security:
        - bearerAuth: []
      tags: ["deleting of a message"]
      operationId: deleteMessage
      summary: deletes the message
      description: User deletes a message
      responses:
        '200':
          description: message deletion action successful
          content:
            plain/text:
              schema:
                description: '200 response message'
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Deletion accepted!"
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{user}/conversations/{conv_id}/messages/{mess_id}/comments:
    description: comments area, where user can comment a message
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: conv_id
        description: receiver id (group or user)
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: mess_id
        description: message id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    put:
      security:
        - bearerAuth: []
      tags: ["comment creation"]
      operationId: commentMessage
      summary: new comment is created
      description: User comments the message defined by the messId
      requestBody:
        description: User gives emoji in input
        content:
          application/json:
            schema:
              type: object
              description: object that contains an emoji in input
              properties:
                emoji:
                  description: emoji in input
                  type: string
                  enum:
                    - "😀"
                    - "😂"
                    - "😍"
                    - "😎"
                    - "😭"
                    - "😡"
                    - "🎉"
                    - "❤️"
                    - "👍"
                    - "🔥"
        required: true
      responses:
        '200':
          description: Comment modify action successful
          content:
            application/json:
              schema:
                type: object
                description: object that contains comment output data
                properties:
                  id:
                    $ref: '#/components/schemas/GeneralId'
                  message_id:
                    $ref: '#/components/schemas/GeneralId'
                  emoji:
                    description: content of the comment, that is a emoji
                    type: string
                    enum:
                      - "😀"
                      - "😂"
                      - "😍"
                      - "😎"
                      - "😭"
                      - "😡"
                      - "🎉"
                      - "❤️"
                      - "👍"
                      - "🔥"
                  sender:
                    $ref: '#/components/schemas/User'
                  conv_id:
                    $ref: '#/components/schemas/GeneralId'
                required: [id, message_id, emoji, sender, conv_id]
        '201':
          description: New comment sending action successful
          content:
            application/json:
              schema:
                type: object
                description: object that contains comment output data
                properties:
                  id:
                    $ref: '#/components/schemas/GeneralId'
                  message_id:
                    $ref: '#/components/schemas/GeneralId'
                  emoji:
                    description: content of the comment, that is a emoji
                    type: string
                    enum:
                      - "😀"
                      - "😂"
                      - "😍"
                      - "😎"
                      - "😭"
                      - "😡"
                      - "🎉"
                      - "❤️"
                      - "👍"
                      - "🔥"
                  sender:
                    $ref: '#/components/schemas/User'
                  conv_id:
                    $ref: '#/components/schemas/GeneralId'
                required: [id, message_id, emoji, sender, conv_id]
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{user}/groups:
    description: user area where he can create group
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    post:
      security:
        - bearerAuth: []
      tags: ["creation of a group"]
      summary: New group is created
      description: User creates a new group with some users
      operationId: groupCreation
      requestBody:
        description: |
          User creates a new group with some users. He gives in input
          a array of users, a name and a photo for the group
        content:
          application/json:
            schema:
              type: object
              description: object that contains an array of user and groupname in input
              properties:
                users:
                  description: array of user ids in input 
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
                  minItems: 1
                  maxItems: 49
                group_name:
                  description: group name in input
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 20
              required: [group_name, users]
        required: true
      responses:
        '201':
          description: New group creaction action successful
          content:
            application/json:
              schema:
                type: object
                description: object that contains group data
                properties:
                  group:
                    $ref: '#/components/schemas/Group'
                  conv_id:
                    $ref: '#/components/schemas/GeneralId'
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{user}/groups/{group_id}:
    description: user area where he can create group
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: group_id
        description: group id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    put:
      security:
        - bearerAuth: []
      tags: ["users group modify"]
      operationId: addToGroup
      summary: Adds users to group
      description: User Adds users to group, defined by the groupId
      requestBody:
        description: User gives an array of usernames in input
        content:
          application/json:
            schema:
              type: object
              description: object that contains an array of usernames in input
              properties:
                users:
                  description: usernames in input
                  type: array
                  items:
                    $ref: '#/components/schemas/Username'
                  minItems: 1
                  maxItems: 49
        required: true
      responses:
        '200':
          description: group users list addition action successful
          content:
            application/json:
              schema:
                type: object
                description: object that contains an array of group members
                properties:
                  members:
                    description: members of the group, unless you
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                    minItems: 1
                    maxItems: 49
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      security:
        - bearerAuth: []
      tags: ["deleting of a group"]
      operationId: leaveGroup
      summary: deletes the group
      description: User leaves the group represented by the groupId
      responses:
        '200':
          description: group deletion action successful
          content:
            plain/text:
              schema:
                description: '200 response message'
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "group leave accepted!"
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...